+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - `$incron::purge_noop` parameter to run purging in `noop` mode.
 - Manage allowed or denied users using `$incron::allowed_users` and `$incron::denied_users`.
+- Manage package version through `$incron::package_version` parameter.
 
 ### Changed
 - Split up purging into a separate `incron::purge` manifest.